Tweak GitHub Workflow Sealights concurrency group
I'm not 100% confident about this, but the theory is that sometimes
the `github.event_name` is actually 'pull_request_target' instead of
'pull_request' and the logic is incorrectly falling back to the
github.ref value, which means my CI check gets cancelled incorrectly.

With this change, whenever there is a pull request number the number
will be used in the concurrency group.

Assisted-by: Chat GPT

Ref: https://issues.redhat.com/browse/EC-1345