Move acceptance test out of Sealights workflow
This is an experiment to see if the current acceptance tests
timeouts have anything to do with Sealights. I suspect not, but
let's try it.

Update: I did see acceptance tests pass in the
.github/workflows/checks-codecov.yaml workflow, so it seems that the
remaining timeout problem is somehow related to Sealights. Let's
disable it to get merges unblocked, and file another story to figure
out why Sealights acceptance tests are timing out.

Ref: https://issues.redhat.com/browse/EC-1387