Improve CLI --help description
This commit aims to create a description that is more informative and
useful to the end user, by explaining the capabilities of the CLI and
its role in the supply chain security processes.

It also includes a blurb that points to contributions page of the team,
where users can find all the relevant info and resources to get
involved.

Assisted by: Cursor (powered by Claude 4 Sonnet)

Ref: https://github.com/conforma/cli/issues/1460
Ref: https://issues.redhat.com/browse/EC-493