Upgrade to latest conftest go module
Notably this is the first version of the conftest module that
uses/supports the v1.x opa module.

It required a little extra manual tweaking to remove the `replace
muzzammil.xyz/jsonc ...` which is not needed now after
https://github.com/open-policy-agent/conftest/pull/1133 was merged.

At this point I'm not sure exactly what the impact will be, but we
have not updated this for a while. Now that we updated opa, let's
update this also.

Edit: There was quite a lot of impact actually, see the other
commits in this PR.

Ref: https://issues.redhat.com/browse/EC-1408
Ref: https://issues.redhat.com/browse/EC-1130