Update ec test source after conftest update
The code in cmd/test/test.go is meant to be an "almost" copy of the
conftest cmd/test/test.go. (It's only functional difference is to
support one additional output format, the "appstudio" format).

To update the file I used this helper to compare the upstream
conftest file with our local copy:
  DIFF_TOOL=vimdiff make conftest-test-cmd-diff

There are some flags changing here, so this could be a breaking
change for users of `ec test`.

As well as the cmd change, there is a snapshot adjustment due to a
new metadata key that Conftest is producing.

FWIW I don't think this is a good strategy long term. Maintaining
this file manually is quite difficult. Maybe it's possible to do it
a different way, e.g. the way it's done in cmd/opa, but still add
support for the additional format. Also I'm not sure if anyone is
using this experimental feature, so we could consider dropping it.
(But let's worry about all that later.)

Ref: https://issues.redhat.com/browse/EC-1408