Fix invalid rego in init policies
We need to use the newer opa v1 syntax. This fixes a failing
acceptance test in features/initialize.feature.

Side-note: The `ec init policy` command is a bit undercooked, and I
don't it's been used much. In this commit I'm just doing the minimum
required to make the tests pass, and resisting the urge to also try
to improve it.

Ref: https://issues.redhat.com/browse/EC-1408