Remove CLI_NAME build arg
It's used in the labels, mainly to make the two Dockerfiles more
easy to diff compare. We're using the same name for everywhere now
so it's no longer needed.

Note that we still have the BUILD_SUFFIX build arg which gets used
in the version, and provides a way distinguish between the two types
of builds.

Ref: https://issues.redhat.com/browse/EC-1383