Fix untrusted rpms-signature-scan task violation
In commit ec1ade4a4f02a8cf812e245287e39d6f6cee27e6 we added a task
ref from the non-standard location, which caused a violation in our
tenant release pipeline.

This should fix it.

In the long term I think all the tasks are going back to the main
tekton-catalog, but in the short term, this is needed so Conforma
considers the task trusted.