Refactor rekor entries creation in acceptance tests
This commit refactors the functions 'CreateAndPushImageSignature' and
'createAndPushAttestationWithPatches' so that they now replicate the
transparency log entry creation on the rekor stub.

The signature and the attestation are now created using the
cosign.TLogUpload function, and stubbing the rekor endpoints that get
called during the tlog entry creation process.
The result is a signature that can be successfully verified using the
'cosign verify' command, and an attestation that has a corresponding
entry in rekor.

This refactor also removed the need to explicitly create rekor entries
in the acceptance tests, since this is now part of the cosign flow.

The acceptance tests using this new rekor flow now reflect more
accuratly the real-world scenario.

Assisted by: Claude Code

Ref: https://issues.redhat.com/browse/EC-1210