Results difference
The results differed between validate image and
the fallback validate image used in validate vsa.
The root cause was when validating the identifier
passed was not a file, the "latest" tag was
appended to the image ref.
This was fixed and the validation code has been
simplified.

Assisted by: Claude Sonnet