chore(deps): Update k8s.io/kubernetes go module
After updating containerd there was a compatibility problem related
to the structured-merge-diff/v4 vs structured-merge-diff/v6 which
caused problems build kubectl using tools/kubectl/go.mod.

This should fix that. I decided to update all the other mod files
similarly so we're using the same version everywhere. Hopefully that
won't introduce any other compatibility problems, but let's see.

Done like this:

    for d in . acceptance tools tools/kubectl; do (cd $d; go get k8s.io/kubernetes && go mod tidy); done
    for d in . acceptance tools tools/kubectl; do (cd $d; go get k8s.io/apiextensions-apiserver && go mod tidy); done
    for d in . acceptance tools tools/kubectl; do (cd $d; go get k8s.io/kubectl && go mod tidy); done

Not sure if that was the optimal way to do it, but I was aiming to
get rid of all the 0.31.0 k8s deps.