Better vsa signing secrets with passwords support
In the PR for https://issues.redhat.com/browse/EC-1586 I began
creating the signing secret using the same kind of GitOps job that
Konflux uses for the Chains signing secret. As a side-effect of
that, we now have a signing key with a password, which I think
doesn't quite work as it should.

Rather than create the secret without a password, let's improve
things a little so we can use the password conveniently.

Note: This would be simpler if we stopped supporting the signing
secret in a file, with the COSIGN_PASSWORD, but I think maybe we
wanna keep that since it might be useful in non-Konflux
environments.

Ref: https://issues.redhat.com/browse/EC-1589